refactor(orchestrator): extract ToolContextBudget from AgentOrchestrator
R2 step 3 (started with the static token-budget cluster, not the plan's
step 1 SPI, mirroring Wave R's "static clusters first" rule): moves
resolveToolContextEstimator/enforceToolContextBudget/findToolExchanges/
tokensOf/sumTokens/sumInt/tokenUsageMap into a new
ai.labs.eddi.modules.llm.impl.orchestration.ToolContextBudget.

Widens LlmTask.resolveModelName and TokenCounterFactory.extractText to
public for the new subpackage. Delegators kept for every hard class-
qualified test/production reference (DEFAULT_MAX_TOOL_CONTEXT_TOKENS,
enforceToolContextBudget, sumTokens, tokenUsageMap, TOKEN_USAGE_FIELDS).

AgentOrchestrator: 2,725 -> 2,542 lines.
